refactored comands to be in modules
